Allow adding a file extension when downloading files
After Effects appears to rely on assets having a particular extension.
For example, to replace an image, it expects the file to have an image
file extension like `.jpg`. If you feed it an image file that does not
have such an extension, `aerender` will fail rendering.

When downloading a remote image, the nexrender downloader typically
saves it to the working directory, named with the basename of the remote
path. This change allows you to optionally specify a file extension in
the job config, which will be appended to the filename when it is
downloaded.